backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0

# Caching
django-redis==5.4.0
redis==4.6.0

# Task Queue
celery==5.3.4
celery[redis]==5.3.4

# Search
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5

# Monitoring
prometheus-client==0.19.0
django-prometheus==2.3.1

# Message Queue
kafka-python==2.0.2

# Object Storage (for media files)
boto3==1.34.10
django-storages==1.14.2

# API Rate Limiting
django-ratelimit==4.1.0

# Database Connection Pooling
psycopg2-pool==1.1

# Debugging and Performance
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Security
django-guardian==2.4.0

#flower
flower==2.0.1